Move adapter implementations into child namespaces
